CCodeBase::UnicodeToHex の第二引数を SSIZE_T に変更

+++ sakura_core/charset/CCesu8.h
@@ -39,7 +39,7 @@ public:
 	}
 	void GetBom(CMemory* pcmemBom) override;																			//!< BOMデータ取得
 // GetEolはCCodeBaseに移動	2010/6/13 Uchi
-	EConvertResult UnicodeToHex(const wchar_t* cSrc, const int iSLen, WCHAR* pDst, const CommonSetting_Statusbar* psStatusbar){			//!< UNICODE → Hex 変換
+	EConvertResult UnicodeToHex(const wchar_t* cSrc, const SSIZE_T iSLen, WCHAR* pDst, const CommonSetting_Statusbar* psStatusbar){			//!< UNICODE → Hex 変換
 		return CUtf8()._UnicodeToHex( cSrc, iSLen, pDst, psStatusbar, true );
 	}
 };
